Ok, big commit. We actually have shit in it now :)
To begin with : The app has an icon!! It looks good, even though it came from a fucking shitty idea.
The first storyboards have been worked upon. As a matter of fact, everything was working out great when i was still working with labels, buttons and segues.
Then the page view controller came and it all went to hell. I started to implement it, but bumped into some issues and since the night is running pretty late and im running pretty drunk, i'll be calling a day.